created map1 component and deleted the map page
